dynamite, dynamite_runtime, nextcloud: move http client to dynamite_runtime
