+++ packages/dynamite/dynamite/lib/src/openapi_builder.dart
@@ -67,48 +67,13 @@ class OpenAPIBuilder implements Builder {
         "import 'package:built_value/json_object.dart';",
         "import 'package:built_value/serializer.dart';",
         "import 'package:built_value/standard_json_plugin.dart';",
-        "import 'package:cookie_jar/cookie_jar.dart';",
         "import 'package:dynamite_runtime/content_string.dart';",
-        "import 'package:universal_io/io.dart';",
+        "import 'package:dynamite_runtime/http_client.dart';",
         '',
-        "export 'package:cookie_jar/cookie_jar.dart';",
+        "export 'package:dynamite_runtime/http_client.dart';",
         '',
         "part '${p.basename(outputId.changeExtension('.g.dart').path)}';",
         '',
-        Extension(
-          (final b) => b
-            ..name = '${prefix}HttpClientResponseBody'
-            ..on = refer('HttpClientResponse')
-            ..methods.addAll([
-              Method(
-                (final b) => b
-                  ..name = 'bodyBytes'
-                  ..returns = refer('Future<Uint8List>')
-                  ..type = MethodType.getter
-                  ..modifier = MethodModifier.async
-                  ..body = const Code(
-                    '''
-                  final chunks = await toList();
-                  if (chunks.isEmpty) {
-                    return Uint8List(0);
-                  }
-                  return Uint8List.fromList(chunks.reduce((final value, final element) => [...value, ...element]));
-                    ''',
-                  ),
-              ),
-              Method(
-                (final b) => b
-                  ..name = 'body'
-                  ..returns = refer('Future<String>')
-                  ..type = MethodType.getter
-                  ..modifier = MethodModifier.async
-                  ..lambda = true
-                  ..body = const Code(
-                    'utf8.decode(await bodyBytes)',
-                  ),
-              ),
-            ]),
-        ).accept(emitter).toString(),
         Class(
           (final b) => b
             ..name = '${prefix}Response'
@@ -116,20 +81,7 @@ class OpenAPIBuilder implements Builder {
               refer('T'),
               refer('U'),
             ])
-            ..fields.addAll([
-              Field(
-                (final b) => b
-                  ..name = 'data'
-                  ..type = refer('T')
-                  ..modifier = FieldModifier.final$,
-              ),
-              Field(
-                (final b) => b
-                  ..name = 'headers'
-                  ..type = refer('U')
-                  ..modifier = FieldModifier.final$,
-              ),
-            ])
+            ..extend = refer('DynamiteResponse<T, U>')
             ..constructors.add(
               Constructor(
                 (final b) => b
@@ -138,7 +90,7 @@ class OpenAPIBuilder implements Builder {
                       (final name) => Parameter(
                         (final b) => b
                           ..name = name
-                          ..toThis = true,
+                          ..toSuper = true,
                       ),
                     ),
                   ),
@@ -159,59 +111,8 @@ class OpenAPIBuilder implements Builder {
         ).accept(emitter).toString(),
         Class(
           (final b) => b
-            ..name = 'RawResponse'
-            ..fields.addAll([
-              Field(
-                (final b) => b
-                  ..name = 'statusCode'
-                  ..type = refer('int')
-                  ..modifier = FieldModifier.final$,
-              ),
-              Field(
-                (final b) => b
-                  ..name = 'headers'
-                  ..type = refer('Map<String, String>')
-                  ..modifier = FieldModifier.final$,
-              ),
-              Field(
-                (final b) => b
-                  ..name = 'body'
-                  ..type = refer('Uint8List')
-                  ..modifier = FieldModifier.final$,
-              ),
-            ])
-            ..constructors.add(
-              Constructor(
-                (final b) => b
-                  ..requiredParameters.addAll(
-                    ['statusCode', 'headers', 'body'].map(
-                      (final name) => Parameter(
-                        (final b) => b
-                          ..name = name
-                          ..toThis = true,
-                      ),
-                    ),
-                  ),
-              ),
-            )
-            ..methods.add(
-              Method(
-                (final b) => b
-                  ..name = 'toString'
-                  ..returns = refer('String')
-                  ..annotations.add(refer('override'))
-                  ..lambda = true
-                  ..body = const Code(
-                    r"'RawResponse(statusCode: $statusCode, headers: $headers, body: ${utf8.decode(body)})'",
-                  ),
-              ),
-            ),
-        ).accept(emitter).toString(),
-        Class(
-          (final b) => b
             ..name = '${prefix}ApiException'
-            ..extend = refer('RawResponse')
-            ..implements.add(refer('Exception'))
+            ..extend = refer('DynamiteApiException')
             ..constructors.addAll(
               [
                 Constructor(
@@ -255,27 +156,6 @@ class OpenAPIBuilder implements Builder {
               ),
             ),
         ).accept(emitter).toString(),
-        if (hasAnySecurity) ...[
-          Class(
-            (final b) => b
-              ..name = '${prefix}Authentication'
-              ..abstract = true
-              ..methods.addAll([
-                Method(
-                  (final b) => b
-                    ..name = 'id'
-                    ..type = MethodType.getter
-                    ..returns = refer('String'),
-                ),
-                Method(
-                  (final b) => b
-                    ..name = 'headers'
-                    ..type = MethodType.getter
-                    ..returns = refer('Map<String, String>'),
-                ),
-              ]),
-          ).accept(emitter).toString(),
-        ],
       ];
 
       if (spec.components?.securitySchemes != null) {
@@ -291,7 +171,7 @@ class OpenAPIBuilder implements Builder {
                         final fields = ['username', 'password'];
                         b
                           ..name = '${prefix}HttpBasicAuthentication'
-                          ..extend = refer('${prefix}Authentication')
+                          ..extend = refer('DynamiteAuthentication')
                           ..constructors.add(
                             Constructor(
                               (final b) => b
@@ -352,7 +232,7 @@ class OpenAPIBuilder implements Builder {
                       (final b) {
                         b
                           ..name = '${prefix}HttpBearerAuthentication'
-                          ..extend = refer('${prefix}Authentication')
+                          ..extend = refer('DynamiteAuthentication')
                           ..constructors.add(
                             Constructor(
                               (final b) => b
@@ -437,42 +317,7 @@ class OpenAPIBuilder implements Builder {
             (final b) {
               if (isRootClient) {
                 b
-                  ..fields.addAll([
-                    Field(
-                      (final b) => b
-                        ..name = 'baseURL'
-                        ..type = refer('String')
-                        ..modifier = FieldModifier.final$,
-                    ),
-                    Field(
-                      (final b) => b
-                        ..name = 'baseHeaders'
-                        ..type = refer('Map<String, String>')
-                        ..modifier = FieldModifier.final$
-                        ..late = true,
-                    ),
-                    Field(
-                      (final b) => b
-                        ..name = 'httpClient'
-                        ..type = refer('HttpClient')
-                        ..modifier = FieldModifier.final$
-                        ..late = true,
-                    ),
-                    Field(
-                      (final b) => b
-                        ..name = 'cookieJar'
-                        ..type = refer('CookieJar?')
-                        ..modifier = FieldModifier.final$,
-                    ),
-                    if (hasAnySecurity) ...[
-                      Field(
-                        (final b) => b
-                          ..name = 'authentications'
-                          ..type = refer('List<${prefix}Authentication>')
-                          ..modifier = FieldModifier.final$,
-                      ),
-                    ],
-                  ])
+                  ..extend = refer('DynamiteClient')
                   ..constructors.add(
                     Constructor(
                       (final b) => b
@@ -480,109 +325,45 @@ class OpenAPIBuilder implements Builder {
                           Parameter(
                             (final b) => b
                               ..name = 'baseURL'
-                              ..toThis = true,
+                              ..toSuper = true,
                           ),
                         )
                         ..optionalParameters.addAll([
                           Parameter(
                             (final b) => b
                               ..name = 'baseHeaders'
-                              ..type = refer('Map<String, String>?')
+                              ..toSuper = true
                               ..named = true,
                           ),
                           Parameter(
                             (final b) => b
                               ..name = 'userAgent'
-                              ..type = refer('String?')
+                              ..toSuper = true
                               ..named = true,
                           ),
                           Parameter(
                             (final b) => b
                               ..name = 'httpClient'
-                              ..type = refer('HttpClient?')
+                              ..toSuper = true
                               ..named = true,
                           ),
                           Parameter(
                             (final b) => b
                               ..name = 'cookieJar'
-                              ..toThis = true
+                              ..toSuper = true
                               ..named = true,
                           ),
                           if (hasAnySecurity) ...[
                             Parameter(
                               (final b) => b
                                 ..name = 'authentications'
-                                ..toThis = true
-                                ..named = true
-                                ..defaultTo = const Code('const []'),
+                                ..toSuper = true
+                                ..named = true,
                             ),
                           ],
-                        ])
-                        ..body = const Code('''
-                        this.baseHeaders = {
-                          ...?baseHeaders,
-                        };
-                        this.httpClient = (httpClient ?? HttpClient())..userAgent = userAgent;
-                      '''),
+                        ]),
                     ),
-                  )
-                  ..methods.addAll([
-                    Method(
-                      (final b) => b
-                        ..name = 'doRequest'
-                        ..returns = refer('Future<RawResponse>')
-                        ..modifier = MethodModifier.async
-                        ..requiredParameters.addAll([
-                          Parameter(
-                            (final b) => b
-                              ..name = 'method'
-                              ..type = refer('String'),
-                          ),
-                          Parameter(
-                            (final b) => b
-                              ..name = 'path'
-                              ..type = refer('String'),
-                          ),
-                          Parameter(
-                            (final b) => b
-                              ..name = 'headers'
-                              ..type = refer('Map<String, String>'),
-                          ),
-                          Parameter(
-                            (final b) => b
-                              ..name = 'body'
-                              ..type = refer('Uint8List?'),
-                          ),
-                        ])
-                        ..body = const Code(r'''
-                        final uri = Uri.parse('$baseURL$path');
-                        final request = await httpClient.openUrl(method, uri);
-                        for (final header in {...baseHeaders, ...headers}.entries) {
-                          request.headers.add(header.key, header.value);
-                        }
-                        if (body != null) {
-                          request.add(body.toList());
-                        }
-                        if (cookieJar != null) {
-                          request.cookies.addAll(await cookieJar!.loadForRequest(uri));
-                        }
-
-                        final response = await request.close();
-                        if (cookieJar != null) {
-                          await cookieJar!.saveFromResponse(uri, response.cookies);
-                        }
-                        final responseHeaders = <String, String>{};
-                        response.headers.forEach((final name, final values) {
-                          responseHeaders[name] = values.last;
-                        });
-                        return RawResponse(
-                          response.statusCode,
-                          responseHeaders,
-                          await response.bodyBytes,
-                        );
-                      '''),
-                    ),
-                  ]);
+                  );
               } else {
                 b
                   ..fields.add(
