+++ packages/dynamite/dynamite_runtime/lib/http_client.dart
@@ -1,0 +1,1 @@
+export 'src/http_client.dart';
