+++ packages/dynamite/dynamite_runtime/pubspec.yaml
@@ -7,6 +7,8 @@ environment:
 
 dependencies:
   built_value: ^8.5.0
+  cookie_jar: ^4.0.5
+  universal_io: ^2.2.2
 
 dev_dependencies:
   build_runner: ^2.4.2
