+++ packages/nextcloud/pubspec.yaml
@@ -8,7 +8,6 @@ environment:
 dependencies:
   built_collection: ^5.1.1
   built_value: ^8.5.0
-  cookie_jar: ^4.0.2
   crypto: ^3.0.3
   crypton: ^2.0.5
   dynamite_runtime:
